Task JDP200902-2 GroupController - wstępna implementacja endpointów
