+++ src/main/java/com/kodilla/ecommercee/domain/Group.java
@@ -1,0 +1,14 @@
+package com.kodilla.ecommercee.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Group {
+    private Long id;
+    private String nameGroup;
+    private String description;
+}
