+++ src/main/resources/application.properties
@@ -1,8 +1,9 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/kodilla_project?serverTimezone=Europe/Warsaw&allowPublicKeyRetrieval=true&useSSL=False
 spring.datasource.username=kodilla_user
 spring.datasource.password=kodilla_password
+spring.main.allow-bean-definition-overriding=true
 
 spring.jpa.database=mysql
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
